Use `Money` in `Asset`, `Liability`, `Policy`.
Over the course of using `Money` as a type, I found multiple issues
wrt to exceptions. Note that these `Attribute` classes often throw
IllegalArgumentException (Runtime exception, unchecked). I followed
this for the new `Attribute`s as well to keep it consistent.

In addition, I dropped all the `stringToAttributeX` in `personAdapter`
as it doesn't really make sense in that, and shifted that
responsibility to `ParserUtil` as it should have been.