Ignore mixed case, double / trailing space in attr
This means "Muhammad Ali" is the same as "muhammad  ali ", and the same
applies for emails, address, phone, policy names, policy insurer name,
asset name, asset type, etc. Only exception is Notes because that is
multiline and there is a case to be made for protecting multiple spaces
(e.g. as an indent)

This resolves bugs of duplicating names through sneaky json edits.